Add Task class
Creates a new object containing the title, description, due date,
and priority of a user's task.

Private fields store the task's ID, and a project name and ID. The
latter are set when the task is added to a specific project, and should
make it easier to find the right project when manipulating the DOM.

Unique IDs are generated using the uniqid package.