Add sort module
Contains a function which takes an array of tasks, and sorts them based
on due date and priority. Due date takes precedence, with the
assumption meeting deadlines is paramount.

For multiple tasks on the same day, task priority is used to help the
user prioritize those which are most important.