Add Project class
Creates a new object containing a name. Private fields store the
project's ID, and a map of tasks.

Methods allow the map to be retrieved and manipulated. Additionally,
the map can be converted to a sorted array. This is used for display
purposes.

Unique IDs are generated using the uniqid package.