Implement 'Add Project' functionality
View class now holds a reference to the currently active project. This
is used to style the sidebar, and select which tasks to render.

Added a new method init to set up the initial state of the page, such
as adding the default projects.

The 'Add Project' button and accompanying dialog is now fully
functional. Adding a project automatically activates said project and
renders a blank page ready to add tasks.