Add defaultProjectId and activeProjectId
Default projects are added automatically via the TodoList constructor.
defaultProjectId keeps track of this project for display purposes.

activeProjectId is used to keep track of the user's currently
selected project. This allows the correct tasks to be loaded, and new
tasks to be added that specific project.

New methods were added for retrieving both defaultProjectId and
activeProjectId. setActiveProjectId allows the active project to be
updated.